backend/requirements.txt
ansible==10.5.0
ansible-core==2.17.5
ansible-runner==2.4.0
Flask==3.0.3
Flask-SocketIO==5.4.1
jsonpatch==1.33
kubernetes==31.0.0
PyYAML==6.0.2
SQLAlchemy==2.0.35
flasgger==0.9.7.1
fastapi==0.115.3
typing==3.7.4.3
pydantic==2.9.2
uvicorn==0.32.0
paramiko==3.5.0
psycopg2-binary==2.9.10
python-jose==3.3.0
python-multipart==0.0.9
passlib==1.7.4
bcrypt==4.0.1
gunicorn===23.0.0
requests
hvac
ldap3==2.9.1
docker